pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scd-edition"
version = "0.1.0"
description = "GUI for editing and visualizing SCD decomposition results"
readme = "README.md"
license = {text = "CC-BY-NC-4.0"}
authors = [
    {name = "Agnese Grison", email = "agnese.grison16@imperial.ac.uk"},
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["EMG", "decomposition", "GUI", "spike editing", "motor units"]

dependencies = [
    "swarm-contrastive-decomposition",  # SCD package from PyPI
    "motor_unit_toolbox @ git+https://github.com/imendezguerra/motor_unit_toolbox.git",
    "numpy>=1.20.0",
    "torch>=1.9.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "pyqtgraph>=0.13.0",
    "PyQt5>=5.15.0",
    "scikit-learn>=1.0.0",
    "mat73>=0.59",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]

[project.scripts]
scd-edition = "scd_app.gui.main_window:main"

[project.urls]
Homepage = "https://github.com/AgneGris/scd-edition"
Repository = "https://github.com/AgneGris/scd-edition"

[tool.setuptools.packages.find]
where = ["src"]
include = ["scd_app*"]

[tool.setuptools.package-data]
scd_app = [
    "gui/style/fonts/*.ttf",
    "resources/loaders_configs/*.yaml",
]